pyproject.toml
[tool.poetry]
name = "intensive"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.3"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
sqlalchemy-utils = "^0.41.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
python-multipart = "^0.0.9"
pydantic = "^2.7.1"
bcrypt = "^4.1.2"
werkzeug = "^3.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
